vrrp: fix persistent FAULT state with use_vmac when interfaces renamed
If an existing base interface of a VMAC is renamed, delete the VMAC
since the configure base interface no longer exists.

When an existing interface is renamed to match the base interface for
a VMAC for a VRRP instance, for IPv6 when the VMAC interface is created
a link local address is added, so clear the NO_ADDRESS fault flag by
calling try_up_instance(). For IPv4 we do not add an address to the VMAC
when it is created, so will wait for notlink notifications of addresses
on the base interface, which can then be added to be VMAC.

Since 4c7a94a ("vrrp: Make VMAC IPv6 link local address mirror parent
interface"), it was not called anymore when a IPv6 link-local was back
on the VMAC interface.

Fixes: 4c7a94a ("vrrp: Make VMAC IPv6 link local address mirror parent interface")

Based on a patch signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>