all: add code to calculate maximum stack usage and use it for no_swap
When a process has no_Swap specified, if we want to ensure that the
stack is resident in memory we need to know the maximum size that it
is likely to grow to.

This commit adds diagnostic code (usually disabled) to report maximum
usage, so that the code can be updated to know, in advance, the maximum
likely stack usage.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>