vrrp: Don't start up VRRP instances before vrrp_startup_delay expires
If a track_script or other tracker was successful before the
vrrp_startup_delay expired, and the VRRP instance priority was 255,
then the vrrp instance was transitioning to master state immediately
rather than waiting for the startup delay to expire.

This commit now makes the VRRP instance wait until the startup delay
expires.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>