vrrp: add vrrp_delay_after_boot global keyword
Global keyword vrrp_startup_delay causes keepalived to always
delay starting up VRRP instances for the configured time, whenever
keepalived starts.

The original purpose of the keyword was to delay keepalived
starting VRRP instances when it is started at boot time (i.e. by
the init process), in order to allow time for the network interfaces
to settle down.

The problem with it is that if keepalived is stopped and subsequently
started again, there is another delay.

vrrp_delay_after_boot will now only create a delay if the system
uptime is below the configured value, and keepalived will wait until
the system uptime reaches the configured value. This means that there
will not be a delay in the VRRP instances starting if keepalived is
stopped and subsequently restarted.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>